[Silver III] Title: 주유소, Time: 396 ms, Memory: 40020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -9,24 +9,24 @@ public class Main {
 
         int N = Integer.parseInt(br.readLine()); // 도시 갯수
 
-        int[] dis= new int[N-1]; // 거리
-        int[] price = new int[N]; // 지역 당 기름가격
+        Long[] dis= new Long[N-1]; // 거리
+        Long[] price = new Long[N]; // 지역 당 기름가격
 
         StringTokenizer st = new StringTokenizer(br.readLine());
         // 거리 입력
         for(int i = 0; i < N-1; i++) {
-            dis[i] = Integer.parseInt(st.nextToken());
+            dis[i] = Long.parseLong(st.nextToken());
         }
 
         //지역 당 기름 가격 입력
         st = new StringTokenizer(br.readLine());
         for(int i = 0; i < N; i++) {
-            price[i] = Integer.parseInt(st.nextToken());
+            price[i] = Long.parseLong(st.nextToken());
         }
 
         // 맨 처음 기름 가격이 최소인지?
-        int min = price[0];
-        int total = 0;
+        long min = price[0];
+        long total = 0;
         for(int i = 0; i < N-1; i++) {
             if(min > price[i]) {
                 min = price[i];
